# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: |
    This API allows management of client contracts and tracks history of employee contract assignments
  title: Simple emlpoyee contract tracker api
  version: 1.0.0
servers:
- url: /
paths:
  /employees:
    get:
      description: Retrieves all employees
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/employee'
                type: array
          description: OK
      summary: Get all employees
      x-accepts: application/json
    post:
      description: Add a new employee with given name and github username
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_employees_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/employee'
          description: OK
      summary: Create new employee
      x-content-type: application/json
      x-accepts: application/json
  /employees/{employeeId}:
    delete:
      parameters:
      - description: Numeric ID of the employee
        explode: false
        in: path
        name: employeeId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      summary: Delete an employee by ID
      x-accepts: application/json
    get:
      parameters:
      - description: Numeric ID of the employee
        explode: false
        in: path
        name: employeeId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/employee'
          description: OK
      summary: Get an employee by ID
      x-accepts: application/json
    put:
      parameters:
      - description: Numeric ID of the employee
        explode: false
        in: path
        name: employeeId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_employees_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/employee'
          description: OK
      summary: Update an employee by ID
      x-content-type: application/json
      x-accepts: application/json
...
components:
...
  schemas:
    employee:
      description: Employee data
      example:
        github: github
        name: name
        id: 0
      properties:
        id:
          type: integer
        name:
          type: string
        github:
          type: string
      type: object
...
    _employees_post_request:
      properties:
        name:
          type: string
        github:
          type: string
      type: object
...